Add tests for external modules being imported.
This emulates a situation where you might import something
into your `addon/routes.js` from another addon that is assumed
to be a dependency of the host app and is not listed in the
engines specific dependencies. In that scenario the tree that we
provide to broccoli-dependency-funnel would not include the module
in question.

The prior implementation (via rollup) did not error in this scenario
(it marked the missing module as "external" and carried on about its
work). mr-dep-walk currently throws an error when this occurs.